LayoutTests/imported/w3c/web-platform-tests/referrer-policy/same-origin/attr-referrer/cross-origin/http-https/img-tag/keep-origin-redirect/cross-origin.http-expected.txt



FAIL The referrer URL is omitted when a
                                 document served over http requires an https
                                 sub-resource via img-tag using the attr-referrer
                                 delivery method with keep-origin-redirect and when
                                 the target request is cross-origin. assert_equals: Reported Referrer URL is 'omitted'. expected (undefined) undefined but got (string) "http://localhost:8800/referrer-policy/same-origin/attr-referrer/cross-origin/http-https/img-tag/keep-origin-redirect/cross-origin.http.html"
FAIL The referrer URL is omitted when a
                                 document served over http requires an https
                                 sub-resource via img-tag using the attr-referrer
                                 delivery method with keep-origin-redirect and when
                                 the target request is cross-origin. (srcdoc iframe inherits parent) assert_equals: Reported Referrer URL is 'omitted'. expected (undefined) undefined but got (string) "http://localhost:8800/referrer-policy/same-origin/attr-referrer/cross-origin/http-https/img-tag/keep-origin-redirect/cross-origin.http.html"
PASS The referrer URL is omitted when a
                                 document served over http requires an https
                                 sub-resource via img-tag using the attr-referrer
                                 delivery method with keep-origin-redirect and when
                                 the target request is cross-origin. (overridden by srcdoc iframe)
